Optimized choose_pair_or_triple using hashmaps to remember already computed pairs; changed Makefile to compile with g++; made changes to several files to make them compilable with g++ and remove any warnings given by g++; updated primer_boundary test outputs as several statistics change when using the optimized version; changed some of the test scripts to correctly give the --log-file argument to newer version of valgrind (3.4)
git-svn-id: http://svn.code.sf.net/p/primer3/code@1056 d7575951-a1bd-4e41-b076-667b2046c51c

+++ primer3/branches/optimizations/src/thal_main.c
@@ -240,7 +240,7 @@ if(a.debug == 0) {
 #endif
 	    exit(-1);
 	 }
-	 a.maxLoop = strtod(argv[i+1], &endptr);
+	 a.maxLoop = (int) (strtod(argv[i+1], &endptr));
 		 
 	 if(a.maxLoop > MAX_LOOP ) {
 	    a.maxLoop = MAX_LOOP;
